Remove bogus/not needed &$… by-reference prefixes from hooks
These & are not only not needed, they make the code more brittle and
harder to read. These parameters are not meant to be passed back to
the caller by reference. The hook handler code is not supposed to
*replace* these objects with something else.

These & are a relic from very old PHP versions and not needed any
more.

Change-Id: I7b9d7ebc023c9f5bbd26dcce0691a7e515937b8b