Unbreak extension for MediaWiki 1.39+
This is a lazy, albeit tested and confirmed functional, patch.
We assume that the actor migration process is complete and we can use the core actor table as it always was meant to be used.
We fill contributors.cn_user_id with actor.actor_user (which is literally the same as user.user_id) and contributors.cn_user_text with actor.actor_name (which is equivalent to user.user_name).
In the future, both the cn_user_id and cn_user_text fields should be dropped and replaced with a cn_actor field to contain just the actor ID.
Given that this extension interacts with the revision table to such a high degree, that will no doubt improve performance and reduce the size of the table altogether.
However, the current unique index will also need changes and an appropriate maintenance script should be written.

Nevertheless, this lazy patch gets the extension up and running and will likely suffice for most sites using the extension for the time being.

Bug: T306900
Change-Id: I3f5310537d44efc5aeaab8dbad91c28ba10c0c08